requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx
spacy
transformers
torch
vaderSentiment
scikit-learn
numpy
textstat